requirements.txt
pygame-ce>=2.5.0
numpy>=1.24.0
matplotlib>=3.0.0
